requirements.txt
fastapi
uvicorn[standard]
langchain==1.2.15
langchain-core==1.2.31
langchain-openai==1.1.14
langsmith==0.7.32
openai==2.32.0
pydantic==2.13.1
python-dotenv==1.2.2
python-multipart
tiktoken==0.12.0
tenacity==9.1.4
httpx==0.28.1